figured out how to dynamically add inputs depending on how many are in the state array from App (this.state.projects.list)
this is a huge discovery. I can finally progress on both projects and experience (jobs). I feel the projects is nearly complete, and experience is just going to be a copy paste clone pretty much.

I feel very proud and happy that I overcame this block which had me puzzled. time to bring it home and wrap up project