projects now has a delete. Projects is finished
I have projects set up exactly how I imagined it would be. Really happy with the result. I could clean up some of the names before copy pasting it to the experience section lol. But really happy how I conditionally render the add option, the delete options, how I dynamically create the different sections of it, and that it could potentially go on for an unlimited amount if NOT for the conditions I mentioned at the beginning of this commit message. The conditions keep the code in code, so only 3 can be made at a time.