got header looking the way I want it to.
header looks good. spent all day trying to relearn css. Had a hell of a time getting the title to be centered with the subtitle directly under it. I finally settled on making a dummy div, and making the parent justify-content:space-between, and assigning the 3 => dummy, name and title to have fixed min-heights. Worked well. Nailed it. Only took 3 hours of attempts haha