lots of work on the css for the input area.
the inputs, pinfo, socials, skills, projects...etc. are looking good. I like the css on a small screen so far. I haven't added color yet, but everything is lined up well which is good. Time to add some color, change around some inputs to be textareas, and to work on the last section of preview