had an issue with my function that checked if education was blank inside main.js. It didn't run on props being updated. This has been fixed using componentDidUpdate.
the issue was that the function was being run statically, and the value was being stored statically. The function wasn't being run again once the props were changed. Now using componentDidUpdate, it now uses the function properly and updates the internal state of main in main.js so we can check if education goes from being blank to being populated.