able to render the projects properly to the screen. They aren't styled yet.
was having a hard time using forEach. I think for reach I have to use map. Once I switched to map i was able to use the 'method' renderprojectselection properly. prior to that, using the forEach method, nothing would render and I was getting a big caught up on the fact that it wasn't rendering with forEach. Happy to be past that.

also I made the map overly complicated by adding a if() statement to check if there were any other items in the list array. But this was not needed. Just running .map to return the renderprojectSection() function is enough. trying to do if(item !== 'undefined') was adding extra work.