I think i'm going to call it a wrap for now.
I would like to add more placeholders, such as having placeholders for social media links, and an example project and experience in the preview. But honestly i'm so over this project I can't be bothered. I would absolutely want to refactor my code for readability and re do the file structure if I were to keep going. As far as making the place holders, Those are things I have already done, so it's just a matter of redoing them.